requirements.txt
pysftp
django
django-jenkins
south
vectorformats
psycopg2
simplejson
openpyxl==1.5.8
xlrd
pysolr
pep8
mox
mock
python-dateutil
django_coverage
django-userena
django-guardian
django-nose
python-memcached
django-memcached
django-tastypie
django-orm-extensions
-e git://github.com/hcarvalhoalves/django-rest-thumbnails.git@6e949d137246e9d3d8180dbc3fa26c512e753b11#egg=django-rest-thumbnails
httplib2
model_mommy==0.8.1
django-waffle
nose-exclude
django-dajaxice
django-dajax
django-bootstrap-toolkit
